backend/requirements.txt
asgiref==3.7.2
frontend @ file:///../frontend/dist/frontend-0.6.tar.gz
Django==4.2.10
django-debug-toolbar==4.2.0
django-filter==23.5
django-request-logging==0.7.5
djangorestframework==3.14.0
pillow==10.2.0
pytz==2023.3.post1
sqlparse==0.4.4
